# api/internal/gen/kotg/models.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version 2.3.0 DO NOT EDIT.
package api

import (
	"time"
)

const (
	BasicAuthScopes  = "basicAuth.Scopes"
	BearerAuthScopes = "bearerAuth.Scopes"
)

// Defines values for Status.
const (
	Expired Status = "expired"
	Failed  Status = "failed"
	Pending Status = "pending"
	Success Status = "success"
)

// CreateSignRequest defines model for CreateSignRequest.
type CreateSignRequest struct {
	HolderId    string  `json:"holder_id"`
	HolderName  string  `json:"holder_name"`
	Issue       []Issue `json:"issue"`
	Notes       *string `json:"notes,omitempty"`
	RequestUser string  `json:"request_user"`
}

// CreateSignResponse defines model for CreateSignResponse.
type CreateSignResponse struct {
	Interfaces *[]NetworkInterface `json:"interfaces,omitempty"`
	RequestId  *string             `json:"request_id,omitempty"`
	Token      *string             `json:"token,omitempty"`
}

// Issue defines model for Issue.
type Issue struct {
	Copy        int     `json:"copy"`
	Description *string `json:"description,omitempty"`
	Number      string  `json:"number"`
}

// NetworkInterface defines model for NetworkInterface.
type NetworkInterface struct {
	Dsn  *string `json:"dsn,omitempty"`
	Ip   *string `json:"ip,omitempty"`
	Name *string `json:"name,omitempty"`
}

// SignResponse defines model for SignResponse.
type SignResponse struct {
	CreatedAt         *time.Time `json:"created_at,omitempty"`
	HolderId          string     `json:"holder_id"`
	HolderName        string     `json:"holder_name"`
	Issue             []Issue    `json:"issue"`
	LocationLatitude  *float32   `json:"location_latitude,omitempty"`
	LocationLongitude *float32   `json:"location_longitude,omitempty"`
	Notes             *string    `json:"notes,omitempty"`
	RequestId         *string    `json:"request_id,omitempty"`
	RequestUser       string     `json:"request_user"`

	// Sign Base64 encoded signing data.
	Sign      *string    `json:"sign,omitempty"`
	SignedAt  *time.Time `json:"signed_at,omitempty"`
	Status    *Status    `json:"status,omitempty"`
	UpdatedAt *time.Time `json:"updated_at,omitempty"`
}

// SignSubmitRequest defines model for SignSubmitRequest.
type SignSubmitRequest struct {
	LocationLatitude  *float32 `json:"location_latitude,omitempty"`
	LocationLongitude *float32 `json:"location_longitude,omitempty"`

	// Sign Base64 encoded signing data.
	Sign string `json:"sign"`
}

// Status defines model for Status.
type Status string

// RequestID defines model for RequestID.
type RequestID = string

// Token defines model for Token.
type Token = string

// XPortierAgent defines model for XPortierAgent.
type XPortierAgent = string

// PostAuthKeyOtgSignCreateParams defines parameters for PostAuthKeyOtgSignCreate.
type PostAuthKeyOtgSignCreateParams struct {
	XPortierAgent XPortierAgent `json:"X-Portier-Agent"`
}

// PostKeyOtgSignParams defines parameters for PostKeyOtgSign.
type PostKeyOtgSignParams struct {
	XPortierAgent XPortierAgent `json:"X-Portier-Agent"`
}

// GetKeyOtgSignRequestIDParams defines parameters for GetKeyOtgSignRequestID.
type GetKeyOtgSignRequestIDParams struct {
	// Token Token to get access for this request
	Token         Token         `form:"token" json:"token"`
	XPortierAgent XPortierAgent `json:"X-Portier-Agent"`
}

// PostKeyOtgSignRequestIDParams defines parameters for PostKeyOtgSignRequestID.
type PostKeyOtgSignRequestIDParams struct {
	// Token Token to get access for this request
	Token Token `form:"token" json:"token"`
}

// GetKeyOtgSignRequestIDQrParams defines parameters for GetKeyOtgSignRequestIDQr.
type GetKeyOtgSignRequestIDQrParams struct {
	// Token Token to get access for this request
	Token         Token         `form:"token" json:"token"`
	XPortierAgent XPortierAgent `json:"X-Portier-Agent"`
}

// PostAuthKeyOtgSignCreateJSONRequestBody defines body for PostAuthKeyOtgSignCreate for application/json ContentType.
type PostAuthKeyOtgSignCreateJSONRequestBody = CreateSignRequest

// PostAuthKeyOtgSignSubmitRequestIDJSONRequestBody defines body for PostAuthKeyOtgSignSubmitRequestID for application/json ContentType.
type PostAuthKeyOtgSignSubmitRequestIDJSONRequestBody = SignSubmitRequest

// PostKeyOtgSignJSONRequestBody defines body for PostKeyOtgSign for application/json ContentType.
type PostKeyOtgSignJSONRequestBody = CreateSignRequest

// PostKeyOtgSignRequestIDJSONRequestBody defines body for PostKeyOtgSignRequestID for application/json ContentType.
type PostKeyOtgSignRequestIDJSONRequestBody = SignSubmitRequest