# database/data/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.27.0

package data

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type Issue struct {
	ID            int32
	SignRequestID string
	Number        string
	Copy          int32
	Description   pgtype.Text
}

type NetworkInterface struct {
	ID            int32
	SignRequestID string
	Name          pgtype.Text
	Ip            pgtype.Text
	Dsn           pgtype.Text
}

type SignRequest struct {
	RequestID   string
	RequestUser string
	HolderID    string
	HolderName  string
	Notes       pgtype.Text
	Status      pgtype.Text
	Token       string
	CreatedAt   pgtype.Timestamptz
	UpdatedAt   pgtype.Timestamptz
	SignedAt    pgtype.Timestamptz
}

type SignSubmission struct {
	ID                int32
	SignRequestID     string
	Sign              []byte
	LocationLatitude  pgtype.Float8
	LocationLongitude pgtype.Float8
	SubmittedAt       pgtype.Timestamptz
}